requirements.txt
prompt_toolkit>=3.0.43
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
